Get CSRF token for bus stops departures
